requirements.txt
json
uuid
email
sentence-transformers
openai
pinecone-client
streamlit
